+++ dart/lib/src/client.dart
@@ -1,4 +1,5 @@
 import 'dart:async';
+import 'dart:math';
 
 import 'package:meta/meta.dart';
 import 'package:sentry/sentry.dart';
@@ -18,6 +19,7 @@ abstract class SentryClient {
   factory SentryClient(SentryOptions options) => createSentryClient(options);
 
   SentryClient.base(this.options, {String origin}) {
+    _random = options.sampleRate == null ? null : Random();
     if (options.transport is NoOpTransport) {
       options.transport = Transport(
         options: options,
@@ -43,6 +45,8 @@ abstract class SentryClient {
   /// * https://docs.sentry.io/learn/context/#capturing-the-user
   User userContext;
 
+  Random _random;
+
   /// Reports an [event] to Sentry.io.
   Future<SentryId> captureEvent(
     SentryEvent event, {
@@ -50,6 +54,11 @@ abstract class SentryClient {
     dynamic hint,
   }) async {
     event = _processEvent(event, eventProcessors: options.eventProcessors);
+
+    // dropped by sampling or event processors
+    if (event == null) {
+      return Future.value(SentryId.empty());
+    }
 
     event = _applyScope(event: event, scope: scope);
 
@@ -100,6 +109,14 @@ abstract class SentryClient {
     dynamic hint,
     List<EventProcessor> eventProcessors,
   }) {
+    if (_sampleRate()) {
+      options.logger(
+        SentryLevel.debug,
+        'Event ${event.eventId.toString()} was dropped due to sampling decision.',
+      );
+      return null;
+    }
+
     for (final processor in eventProcessors) {
       try {
         event = processor(event, hint);
@@ -159,5 +176,12 @@ abstract class SentryClient {
       }
     }
     return event;
+  }
+
+  bool _sampleRate() {
+    if (options.sampleRate != null && _random != null) {
+      return (options.sampleRate < _random.nextDouble());
+    }
+    return true;
   }
 }